+++ src/main/java/com/dryadandnaiad/sethlans/services/database/BlenderBinaryDatabaseServiceImpl.java
@@ -46,7 +46,7 @@ public class BlenderBinaryDatabaseServiceImpl implements BlenderBinaryDatabaseSe
     }
 
     @Override
-    public BlenderBinary getById(Integer id) {
+    public BlenderBinary getById(Long id) {
         return blenderBinaryRepository.findOne(id);
     }
 
@@ -56,7 +56,7 @@ public class BlenderBinaryDatabaseServiceImpl implements BlenderBinaryDatabaseSe
     }
 
     @Override
-    public void delete(Integer id) {
+    public void delete(Long id) {
         BlenderBinary blenderBinary = blenderBinaryRepository.findOne(id);
         blenderBinaryRepository.delete(blenderBinary);
