+++ src/main/java/com/dryadandnaiad/sethlans/services/database/BlenderRenderTaskDatabaseServiceImpl.java
@@ -50,7 +50,7 @@ public class BlenderRenderTaskDatabaseServiceImpl implements BlenderRenderTaskDa
     }
 
     @Override
-    public BlenderRenderTask getById(Integer id) {
+    public BlenderRenderTask getById(Long id) {
         return blenderRenderTaskRepository.findOne(id);
     }
 
@@ -60,7 +60,7 @@ public class BlenderRenderTaskDatabaseServiceImpl implements BlenderRenderTaskDa
     }
 
     @Override
-    public void delete(Integer id) {
+    public void delete(Long id) {
         BlenderRenderTask blenderRenderTask = blenderRenderTaskRepository.findOne(id);
         blenderRenderTaskRepository.delete(blenderRenderTask);
