+++ src/main/java/com/dryadandnaiad/sethlans/services/database/CRUDService.java
@@ -30,9 +30,9 @@ import java.util.List;
 public interface CRUDService<T> {
     List<T> listAll();
 
-    T getById(Integer id);
+    T getById(Long id);
 
     T saveOrUpdate(T domainObject);
 
-    void delete(Integer id);
+    void delete(Long id);
 }
