+++ src/main/java/com/dryadandnaiad/sethlans/services/database/NotificationDatabaseServiceImpl.java
@@ -49,7 +49,7 @@ public class NotificationDatabaseServiceImpl implements NotificationDatabaseServ
     }
 
     @Override
-    public SethlansNotification getById(Integer id) {
+    public SethlansNotification getById(Long id) {
         return notificationRepository.findOne(id);
     }
 
@@ -76,7 +76,7 @@ public class NotificationDatabaseServiceImpl implements NotificationDatabaseServ
     }
 
     @Override
-    public void delete(Integer id) {
+    public void delete(Long id) {
         SethlansNotification sethlansNotification = notificationRepository.findOne(id);
         notificationRepository.delete(sethlansNotification);
     }
