+++ src/main/java/com/dryadandnaiad/sethlans/services/database/SethlansNodeDatabaseServiceImpl.java
@@ -51,7 +51,7 @@ public class SethlansNodeDatabaseServiceImpl implements SethlansNodeDatabaseServ
     }
 
     @Override
-    public SethlansNode getById(Integer id) {
+    public SethlansNode getById(Long id) {
         return nodeRepository.findOne(id);
     }
 
@@ -61,7 +61,7 @@ public class SethlansNodeDatabaseServiceImpl implements SethlansNodeDatabaseServ
     }
 
     @Override
-    public void delete(Integer id) {
+    public void delete(Long id) {
         SethlansNode sethlansNode = nodeRepository.findOne(id);
         String connectionURL = "https://" + sethlansNode.getIpAddress() + ":" + sethlansNode.getNetworkPort() + "/api/nodeactivate/removal";
         String params = "connection_uuid=" + sethlansNode.getConnection_uuid();
