+++ src/main/java/com/dryadandnaiad/sethlans/services/database/SethlansServerDatabaseServiceImpl.java
@@ -51,7 +51,7 @@ public class SethlansServerDatabaseServiceImpl implements SethlansServerDatabase
     }
 
     @Override
-    public SethlansServer getById(Integer id) {
+    public SethlansServer getById(Long id) {
         return serverRepository.findOne(id);
     }
 
@@ -61,7 +61,7 @@ public class SethlansServerDatabaseServiceImpl implements SethlansServerDatabase
     }
 
     @Override
-    public void delete(Integer id) {
+    public void delete(Long id) {
         SethlansServer sethlansServer = serverRepository.findOne(id);
         this.applicationEventPublisher.publishEvent(new SethlansEvent(this, sethlansServer.getConnection_uuid() + "-" + NotificationOrigin.ACTIVATION_REQUEST, false));
         serverRepository.delete(sethlansServer);
